entirely replace armv7 operands
Operands was an enum covering all forms of operands and data for each
operand for all instructions. this is hard to iterate, and hard to work
with when a single operand is the one of interest, so it's now replaced
by an array of `Operand` enum instances like other architectures.

in the course of this change, several forms of decoding are broken,
while adding support for some earlier-unsupported multiplies and
  instructions like msr and clz

also clearly note which document the comments mentioning page
numbers/figures is referencing